refactor(answer): add prompt tools, extract `pipeline_related_quesiton` for related question extraction
